backend/requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
pyyaml>=6.0
xgboost>=2.0
scikit-learn>=1.3
shap>=0.44
fastapi>=0.110
uvicorn>=0.27
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0
loguru>=0.7
groq>=0.9
google-generativeai>=0.8
sentence-transformers>=2.7
faiss-cpu>=1.7
pytest>=8.0
pytest-asyncio>=0.23
aiosqlite>=0.20
